scope ExpoFcmMessaging service for the client
