+++ android/expoview/src/main/java/host/exp/exponent/fcm/ExpoFcmMessagingService.java
@@ -5,19 +5,12 @@ import com.google.firebase.messaging.RemoteMessage;
 import org.json.JSONException;
 import org.json.JSONObject;
 
-import java.util.Map;
-
 import androidx.annotation.NonNull;
 import expo.modules.notifications.FirebaseListenerService;
-import expo.modules.notifications.notifications.model.NotificationContent;
-import expo.modules.notifications.notifications.model.NotificationRequest;
-import expo.modules.notifications.notifications.model.triggers.FirebaseNotificationTrigger;
 import host.exp.exponent.ABIVersion;
 import host.exp.exponent.Constants;
 import host.exp.exponent.analytics.EXL;
-import host.exp.exponent.kernel.ExperienceId;
 import host.exp.exponent.notifications.PushNotificationHelper;
-import host.exp.exponent.notifications.model.ScopedNotificationRequest;
 import host.exp.exponent.storage.ExperienceDBObject;
 import host.exp.exponent.storage.ExponentDB;
 
@@ -80,19 +73,5 @@ public class ExpoFcmMessagingService extends FirebaseListenerService {
 
   private void dispatchToLegacyNotificationModule(RemoteMessage remoteMessage) {
     PushNotificationHelper.getInstance().onMessageReceived(this, remoteMessage.getData().get("experienceId"), remoteMessage.getData().get("channelId"), remoteMessage.getData().get("message"), remoteMessage.getData().get("body"), remoteMessage.getData().get("title"), remoteMessage.getData().get("categoryId"));
-  }
-
-  @NonNull
-  @Override
-  protected NotificationRequest createNotificationRequest(@NonNull String identifier, @NonNull NotificationContent content, FirebaseNotificationTrigger notificationTrigger) {
-    ExperienceId experienceId;
-    Map<String, String> data = notificationTrigger.getRemoteMessage().getData();
-    if (!data.containsKey("experienceId")) {
-      experienceId = null;
-    } else {
-      experienceId = ExperienceId.create(data.get("experienceId"));
-    }
-    String experienceIdString = experienceId == null ? null : experienceId.get();
-    return new ScopedNotificationRequest(identifier, content, notificationTrigger, experienceIdString);
   }
 }
